requirements.txt
Django==3.2
psycopg2-binary>=2.8
djangorestframework==3.13.1
django-cors-headers==3.13.0
celery==5.1.2
redis==3.5.3

pygsheets==2.0.5
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4


